+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - updated Addressable::Template to use RFC 6570, level 4
 - fixed compatibility problems with some versions of Ruby
 - moved unicode tables into a data file for performance reasons
+- removing support for multiple query value notations
 
 # Addressable 2.2.8
 - fixed issues with dot segment removal code